Added deploy script, jenkins will execute it after successfully build.

+++ deploy.sh
@@ -3,7 +3,14 @@
 echo "build successfully, deploying to the server..."
 cd /var/lib/jenkins/workspace/whattocook
 echo "kill the existing process"
-ps -ef|grep -e root|grep -e SNAPSHOT|awk '{print $2}'|xargs sudo kill -9
+
+pid=`ps -ef|grep -e root|grep -e SNAPSHOT|awk '{print $2}'`
+
+if [ ! -n $pid ];
+then
+  echo "no process exists... move on to mvn compiling"
+else
+  ps -ef|grep -e root|grep -e SNAPSHOT|awk '{print $2}'|xargs sudo kill -9
 echo "process killed"
 echo "generating new target..."
 mvn clean package -Dmaven.test.skip=true
